+++ nixos-modules/microvm/options.nix
@@ -415,6 +415,20 @@ in
             description = "Virtiofs caching policy for the file system, ignored when 9p is used";
             default = "auto";
           };
+          posixAcl = mkOption {
+            type = bool;
+            default = true;
+            description = ''
+              Pass `--posix-acl --xattr` to virtiofsd. Disable when using
+              `--translate-uid`/`--translate-gid` (asserted to be mutually exclusive
+              with `--posix-acl`, see asserts.nix).
+            '';
+          };
+          extraArgs = mkOption {
+            type = listOf str;
+            default = [];
+            description = "Extra arguments passed to virtiofsd for this share.";
+          };
         };
       }));
     };
@@ -615,6 +629,23 @@ in
         default = "${hostName}-gpu.sock";
         description = ''
           Path of vhost-user socket
+        '';
+      };
+
+      vulkan = mkOption {
+        type = with types; nullOr (enum [ "venus" "drm_native_context" ]);
+        default = null;
+        description = ''
+          QEMU Vulkan translation protocol to use for enabling virtio-gpu Vulkan support.
+        '';
+      };
+
+      hostmem = mkOption {
+        type = types.str;
+        default = "256M";
+        description = ''
+          Size of the QEMU virtio-gpu host memory window, typically between 256M and 8G.
+          The value must include a unit suffix (M or G).
         '';
       };
     };